Refactor: Move mock test to mocks folder
