+++ test/core/test/mocks/mocked-no-mocks-same.test.ts
@@ -1,7 +1,7 @@
 import { expect, test, vi } from 'vitest'
-import { mockedA } from '../src/mockedA'
+import { mockedA } from '../../src/mockedA'
 
-vi.mock('../src/mockedA.ts')
+vi.mock('../../src/mockedA.ts')
 
 test('testing mocking module without __mocks__ - suites dont conflict', () => {
   mockedA()
