+++ test/core/test/mocks/mocked.test.ts
@@ -1,9 +1,9 @@
 import { assert, expect, test, vi, vitest } from 'vitest'
-import { two } from '../src/submodule'
-import * as mocked from '../src/mockedA'
-import { mockedB } from '../src/mockedB'
+import { two } from '../../src/submodule'
+import * as mocked from '../../src/mockedA'
+import { mockedB } from '../../src/mockedB'
 
-vitest.mock('../src/submodule')
+vitest.mock('../../src/submodule')
 
 test('submodule is mocked to return "two" as 3', () => {
   assert.equal(3, two)
